Adding property data api from launch-toolkit for property data. Added property checks to credentials
