+++ packages/ui-plugins/live-activities/src/hooks/useProfile.ts
@@ -1,5 +1,4 @@
 import {
-  EnvironmentMap,
   useEnvironmentValue,
   useImsAccessToken,
   useImsOrg,
@@ -7,140 +6,12 @@ import {
 } from '@assurance/plugin-bridge-provider';
 import { useQuery } from '@tanstack/react-query';
 import { useECID } from './useClientInfo';
-
-interface Identity {
-  namespace: {
-    code: string;
-  };
-  id: string;
-}
-
-interface PushNotificationDetail {
-  denylisted: boolean;
-  token: string;
-  identity: Identity;
-  platform: string;
-  appID: string;
-}
-
-interface ExtSourceSystemAudit {
-  lastUpdatedDate: string;
-}
-
-interface ConsentValue {
-  val: string;
-}
-
-interface MarketingConsent {
-  preferred: string;
-  push: ConsentValue;
-}
-
-interface ConsentDetails {
-  marketing: MarketingConsent;
-}
-
-interface IdSpecificConsents {
-  ECID: {
-    [ecid: string]: ConsentDetails;
-  };
-}
-
-interface ConsentMetadata {
-  time: string;
-}
-
-interface Consents {
-  metadata: ConsentMetadata;
-  idSpecific: IdSpecificConsents;
-}
-
-interface IdentityMapEntry {
-  id: string;
-}
-
-interface IdentityMap {
-  ecid: IdentityMapEntry[];
-}
-
-interface UserActivityRegion {
-  captureTimestamp: string;
-}
-
-interface UserActivityRegions {
-  [region: string]: UserActivityRegion;
-}
-
-interface ProfileEntity {
-  pushNotificationDetails: PushNotificationDetail[];
-  extSourceSystemAudit: ExtSourceSystemAudit;
-  consents: Consents;
-  identityMap: IdentityMap;
-  userActivityRegions: UserActivityRegions;
-  consentPoliciesIDMap: Record<string, unknown>;
-}
-
-interface MergePolicy {
-  id: string;
-}
-
-interface ProfileApiResponse {
-  entityId: string;
-  mergePolicy: MergePolicy;
-  sources: string[];
-  tags: string[];
-  identityGraph: string[];
-  entity: ProfileEntity;
-  lastModifiedAt: string;
-}
-
-interface ProfileParams {
-  baseUrl: string;
-  ecid: string;
-  org: string;
-  sandbox: string;
-  token: string;
-}
-
-const getProfile = async ({
-  baseUrl,
-  ecid,
-  org,
-  sandbox,
-  token
-}: ProfileParams): Promise<ProfileApiResponse | null> => {
-  const response = await fetch(
-    `${baseUrl}/data/core/ups/access/entities?entityId=${ecid}&entityIdNS=ECID&schema.name=_xdm.context.profile&sandbox=${sandbox}`,
-    {
-      headers: {
-        Authorization: `Bearer ${token}`,
-        'x-api-key': 'Activation-DTM',
-        'x-gw-ims-org-id': org
-      }
-    }
-  );
-  const data: ProfileApiResponse = await response.json();
-  if (!Object.keys(data).length) {
-    return null;
-  }
-
-  // console.log('data here', Object.values(data)[0]);
-
-  return Object.values(data)[0]; 
-};
-
-const baseUrls: EnvironmentMap<string> = {
-  local: 'https://platform.adobe.io',
-  dev: 'https://platform-stage.adobe.io',
-  qa: 'https://platform-stage.adobe.io',
-  stage: 'https://platform.adobe.io',
-  prod: 'https://platform.adobe.io'
-};
+import { getProfile, getProfileBaseUrl } from '../api/profile';
 
 function useProfile() {
   const ecid = useECID();
   const sandbox = useSandbox();
-  const baseUrl = useEnvironmentValue(baseUrls);
+  const baseUrl = getProfileBaseUrl(useEnvironmentValue);
 
   const token = useImsAccessToken();
   const org = useImsOrg();
@@ -148,7 +19,6 @@ function useProfile() {
   return useQuery({
     queryKey: ['profile', ecid, org, token],
     queryFn: () => {
-      // console.log(token, org, ecid, sandbox, 'token, org, ecid, sandbox QUERRYYYYYYYYY');
       if (!token || !org || !ecid || !sandbox) {
         console.error('Token, org, ecid, and sandbox are required', token, org, ecid, sandbox);
         return null;
